+++ labtasker/server/database.py
@@ -257,7 +257,7 @@ class DBService:
         args: Optional[Dict[str, Any]] = None,
         metadata: Optional[Dict[str, Any]] = None,
         cmd: Optional[Union[str, List[str]]] = None,
-        heartbeat_timeout: Optional[int] = None,
+        heartbeat_timeout: Optional[float] = None,
         task_timeout: Optional[
             int
         ] = None,  # Maximum time in seconds for task execution
@@ -467,7 +467,7 @@ class DBService:
         queue_id: str,
         worker_id: Optional[str] = None,
         eta_max: Optional[str] = None,
-        heartbeat_timeout: Optional[int] = None,
+        heartbeat_timeout: Optional[float] = None,
         start_heartbeat: bool = True,
         required_fields: Optional[Dict[str, Any]] = None,
         extra_filter: Optional[Dict[str, Any]] = None,
@@ -484,7 +484,7 @@ class DBService:
             queue_id (str): The id of the queue to fetch the task from.
             worker_id (str, optional): The ID of the worker to assign the task to.
             eta_max (str, optional): The optional task execution timeout override. Recommended using when start_heartbeat is False.
-            heartbeat_timeout (int, optional): The optional heartbeat timeout interval in seconds.
+            heartbeat_timeout (float, optional): The optional heartbeat timeout interval in seconds.
             start_heartbeat (bool): Whether to start heartbeat.
             required_fields (dict, optional): Which fields are required.
             extra_filter (Dict[str, Any], optional): Additional filter criteria for the task.
